# openapi.yaml
openapi: 3.0.3
info:
  title: Cowork Booking Access API
  version: contract-v1-draft
  description: |
    Provider Access, consumer Purchase. State proposed (M2 draft).
    Contract text: CONTRACT.md. Rules AXS-R01 to AXS-R19.
    Natural key booking_reference for issue and revoke (no Idempotency-Key header).
    Errors: {"error": {"code": "...", "message": "..."}}.
    The lock is mocked: a stored grant or a checked_in state is not proof a door opened.
servers:
  - url: http://access:8000
    description: Internal URL used by Purchase (ACCESS_INTERNAL_URL)
  - url: http://localhost:8003
    description: Public URL for the browser and the kiosk (PUBLIC_URL)
tags:
  - name: api
    description: JSON API, bearer ACCESS_API_TOKEN, Purchase only
  - name: e-ticket
    description: View-only bearer link for the Member's browser
  - name: kiosk
    description: Staff check-in page, HTTP Basic
  - name: ops
    description: Health and the test clock
paths:
  /grants:
    post:
      tags: [api]
      operationId: issueGrant
      summary: Issue the grant for one confirmed booking (repeat returns the stored grant)
      description: |
        Order of checks: token (401), fields (400), then by booking_reference:
        revoked grant or tombstone (200 revoked, nothing issued), existing grant
        (200 unchanged, body ignored), new grant (201).
        Rules AXS-R01, AXS-R02, AXS-R03, AXS-R04, AXS-R05, AXS-R06.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantCreate'
            examples:
              paid_booking:
                value:
                  booking_reference: BK-7KQ2M9
                  member_ref: '17'
                  space_id: 1
                  space_name: Meeting Room A
                  valid_from: '2026-10-07T09:00:00+07:00'
                  valid_until: '2026-10-07T10:30:00+07:00'
              plan_booking_no_payment:
                summary: Coverage skips collection; same request shape, no price or coverage
                value:
                  booking_reference: BK-3MZ8QT
                  member_ref: '18'
                  space_id: 1
                  space_name: Meeting Room A
                  valid_from: '2026-10-07T10:30:00+07:00'
                  valid_until: '2026-10-07T11:30:00+07:00'
              naive_window:
                summary: Invalid on purpose, valid_from has no offset; answers 400 invalid_request
                value:
                  booking_reference: BK-7KQ2M9
                  member_ref: '17'
                  space_id: 1
                  space_name: Meeting Room A
                  valid_from: '2026-10-07T09:00:00'
                  valid_until: '2026-10-07T10:30:00+07:00'
      responses:
        '201':
          description: New grant, status issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
              examples:
                issued:
                  $ref: '#/components/examples/GrantIssued'
                plan_booking:
                  $ref: '#/components/examples/GrantPlanIssued'
        '200':
          description: Repeat (the stored grant, unchanged) or a revoked reference (status revoked, nothing issued)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
              examples:
                repeat:
                  $ref: '#/components/examples/GrantIssued'
                tombstone:
                  $ref: '#/components/examples/GrantTombstone'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /grants/{booking_reference}:
    get:
      tags: [api]
      operationId: getGrant
      summary: Read one grant with its derived condition (Purchase does not call it in v1)
      description: Rules AXS-R02, AXS-R04, AXS-R05, AXS-R08, AXS-R16.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingReference'
      responses:
        '200':
          description: A grant or a tombstone (a tombstone is 200 revoked, never 404)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
              examples:
                issued:
                  $ref: '#/components/examples/GrantIssued'
                no_show:
                  $ref: '#/components/examples/GrantNoShow'
                revoked:
                  $ref: '#/components/examples/GrantRevoked'
                tombstone:
                  $ref: '#/components/examples/GrantTombstone'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No grant and no tombstone for the reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: not_found
                  message: grant not found
  /grants/{booking_reference}/revoke:
    post:
      tags: [api]
      operationId: revokeGrant
      summary: Revoke the grant (idempotent; an unknown reference gets a revoked tombstone)
      description: |
        No request body. Revoke changes nothing outside Access and implies no refund.
        Rules AXS-R02, AXS-R04, AXS-R16, AXS-R17.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingReference'
      responses:
        '200':
          description: status revoked; a repeat changes nothing (revoked_at unchanged)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
              examples:
                revoked:
                  $ref: '#/components/examples/GrantRevoked'
                tombstone:
                  $ref: '#/components/examples/GrantTombstone'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /t/{ticket_token}:
    get:
      tags: [e-ticket]
      operationId: showTicket
      security: []
      summary: E-ticket (view-only bearer link, no login)
      description: |
        Markers: element data-ticket-code="H7K3-9QXA" (the large code) and
        element data-status="issued" (stored state: issued, checked_in or revoked).
        QR (inline SVG, segno) encodes exactly the code. Referrer-Policy no-referrer.
        Rules AXS-R05, AXS-R07, AXS-R08, AXS-R09, AXS-R10.
      parameters:
        - name: ticket_token
          in: path
          required: true
          description: Issued tokens are secrets.token_urlsafe(16), 22 characters matching ^[A-Za-z0-9_-]{22}$; any unknown value, well-formed or not, gets 404, never 400 (no pattern here on purpose)
          schema:
            type: string
          example: q3Vt9XbLk2Rm8PzW4nHc7A
      responses:
        '200':
          description: The e-ticket; a CANCELLED overlay when revoked
          headers:
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
          content:
            text/html:
              schema:
                type: string
              example: '<span data-status="issued">Issued</span><p data-ticket-code="H7K3-9QXA">H7K3-9QXA</p>'
        '404':
          description: Unknown token; no hint whether any ticket exists
          content:
            text/html:
              schema:
                type: string
  /checkin:
    get:
      tags: [kiosk]
      operationId: showKiosk
      summary: Kiosk page with the room list, the code input, the last result and the last 10 scans
      description: |
        After a scan the flashed result element is data-result="ok" (or not_open_yet, closed,
        revoked, wrong_room, unknown_code), shown once; only that element carries data-result.
        The kiosk carries data-selected-space-id; each of the last 10 scan rows carries
        data-scan-result and data-scan-last4, newest first by scan time, then scan id.
        Each room is labelled with its space_name and room number, "Meeting Room A (room 1)", listed by space_id ascending.
        With no grant stored: "No rooms yet: a room appears after its first ticket is issued"
        and no code input. Rules AXS-R11, AXS-R14, AXS-R15.
      security:
        - basicAuth: []
      responses:
        '200':
          description: Kiosk page
          content:
            text/html:
              schema:
                type: string
              example: '<div data-result="ok">Door unlocked (mock)</div>'
        '401':
          $ref: '#/components/responses/BasicChallenge'
    post:
      tags: [kiosk]
      operationId: postKiosk
      summary: Select the room (space_id, no code) or scan a code at the selected room
      description: |
        Two POSTs: first space_id alone selects the room (stored in access_session), then
        code alone scans at that room with the same cookie. A POST with code scans at the
        room stored in access_session; a space_id sent with a code is ignored, so a cross-site
        scan (no SameSite=Lax cookie) has no room and is refused ("Select the room first"). A
        cross-site room change (space_id, no code) is possible and accepted: the browser resends
        the cached Basic credentials (ADR-0016, AXS-R11 row 10). With no room
        selected the flash is "Select the room first" and no scan is stored. A space_id not in
        the room list gets "Unknown room" and the selection does not change. A code blank after
        normalisation gets "Enter a code" and no scan is stored. Order: unknown_code, revoked,
        wrong_room, then the window.
        Rules AXS-R11, AXS-R12, AXS-R13, AXS-R14, AXS-R15, AXS-R16.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                space_id:
                  type: integer
                  description: Selects the room when sent without code; ignored when code is present
                  example: 1
                code:
                  type: string
                  description: Upper-cased, spaces and hyphens stripped before lookup
                  example: H7K3-9QXA
            examples:
              select_room:
                value:
                  space_id: 1
              scan:
                value:
                  code: H7K3-9QXA
      responses:
        '303':
          description: Redirect to /checkin, which shows the result once
          headers:
            Location:
              schema:
                type: string
              example: /checkin
        '401':
          $ref: '#/components/responses/BasicChallenge'
  /health:
    get:
      tags: [ops]
      operationId: health
      security: []
      summary: Liveness with the deployed revision
      responses:
        '200':
          description: Database reachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
              example:
                status: ok
                revision: 3f9c2e1
        '503':
          description: Database unreachable. /health keeps the seed shape, not the standard error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
              example:
                status: error
                error: database unreachable
  /_test/clock:
    post:
      tags: [ops]
      operationId: setTestClock
      security: []
      summary: Set or clear the test clock (404 unless TEST_CLOCK_ENABLED=true)
      description: Rule AXS-R19.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Clock'
            examples:
              set:
                value:
                  now: '2026-10-07T09:00:00+07:00'
              clear:
                value:
                  now: null
      responses:
        '200':
          description: Override stored (or cleared with null)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clock'
              example:
                now: '2026-10-07T09:00:00+07:00'
        '400':
          description: A now that is not ISO 8601 with an offset; nothing stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: TEST_CLOCK_ENABLED is not exactly "true"; nothing stored. With the flag off, clock.now() never reads test_clock
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Authorization: Bearer <ACCESS_API_TOKEN>. Checked before validation; 401 otherwise.'
    basicAuth:
      type: http
      scheme: basic
      description: User "staff", password STAFF_PASSWORD (only the password is checked).
  parameters:
    BookingReference:
      name: booking_reference
      in: path
      required: true
      schema:
        type: string
        pattern: '^BK-[23456789ABCDEFGHJKMNPQRSTVWXYZ]{6}$'
      example: BK-7KQ2M9
  responses:
    InvalidRequest:
      description: A field or the path reference is invalid; the message names it; nothing stored
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            naive_time:
              value:
                error:
                  code: invalid_request
                  message: valid_from needs a UTC offset
            empty_window:
              value:
                error:
                  code: invalid_request
                  message: valid_until must be after valid_from
            space_id_out_of_range:
              value:
                error:
                  code: invalid_request
                  message: space_id is out of range
    Unauthorized:
      description: No token, a wrong token or the wrong scheme; nothing changes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: unauthorized
              message: unauthorized
    BasicChallenge:
      description: Missing or wrong Staff password; no scan stored (gunicorn's access log keeps the 401 line)
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Basic realm="checkin"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, unauthorized, not_found]
            message:
              type: string
    GrantCreate:
      type: object
      required: [booking_reference, member_ref, space_id, space_name, valid_from, valid_until]
      properties:
        booking_reference:
          type: string
          pattern: '^BK-[23456789ABCDEFGHJKMNPQRSTVWXYZ]{6}$'
        member_ref:
          type: string
          minLength: 1
          description: Opaque Member id as text; never an email; never returned
        space_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
        space_name:
          type: string
          minLength: 1
        valid_from:
          type: string
          format: date-time
          description: ISO 8601 with an offset; the booking start
        valid_until:
          type: string
          format: date-time
          description: ISO 8601 with an offset, after valid_from; the booking end (exclusive)
    Grant:
      type: object
      required: [grant_id, booking_reference, status, condition, ticket_code, ticket_url, space_id, space_name, valid_from, valid_until, revoked_at]
      properties:
        grant_id:
          type: string
          pattern: '^gr_'
        booking_reference:
          type: string
        status:
          type: string
          enum: [issued, checked_in, revoked]
        condition:
          type: string
          enum: [not_yet_valid, expired, no_show, null]
          nullable: true
          description: Derived on read, never stored; always null when revoked
        ticket_code:
          type: string
          pattern: '^[23456789ABCDEFGHJKMNPQRSTVWXYZ]{4}-[23456789ABCDEFGHJKMNPQRSTVWXYZ]{4}$'
          nullable: true
        ticket_url:
          type: string
          format: uri
          nullable: true
        space_id:
          type: integer
          nullable: true
        space_name:
          type: string
          nullable: true
        valid_from:
          type: string
          format: date-time
          nullable: true
        valid_until:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, error]
        revision:
          type: string
        error:
          type: string
    Clock:
      type: object
      required: [now]
      properties:
        now:
          type: string
          format: date-time
          nullable: true
  examples:
    GrantIssued:
      value:
        grant_id: gr_8Fv2Lq9Xk3Tw7Mz1Rb5Nc4
        booking_reference: BK-7KQ2M9
        status: issued
        condition: not_yet_valid
        ticket_code: H7K3-9QXA
        ticket_url: http://localhost:8003/t/q3Vt9XbLk2Rm8PzW4nHc7A
        space_id: 1
        space_name: Meeting Room A
        valid_from: '2026-10-07T09:00:00+07:00'
        valid_until: '2026-10-07T10:30:00+07:00'
        revoked_at: null
    GrantPlanIssued:
      value:
        grant_id: gr_6Wc3Nv8Jt2Xq5Pb9Hm4Rd7
        booking_reference: BK-3MZ8QT
        status: issued
        condition: not_yet_valid
        ticket_code: M4TR-8WCE
        ticket_url: http://localhost:8003/t/Zk7Pq2Wn9Tb4Xm6Rc3Lv8H
        space_id: 1
        space_name: Meeting Room A
        valid_from: '2026-10-07T10:30:00+07:00'
        valid_until: '2026-10-07T11:30:00+07:00'
        revoked_at: null
    GrantNoShow:
      summary: Clock 2026-10-07 10:30, never scanned
      value:
        grant_id: gr_8Fv2Lq9Xk3Tw7Mz1Rb5Nc4
        booking_reference: BK-7KQ2M9
        status: issued
        condition: no_show
        ticket_code: H7K3-9QXA
        ticket_url: http://localhost:8003/t/q3Vt9XbLk2Rm8PzW4nHc7A
        space_id: 1
        space_name: Meeting Room A
        valid_from: '2026-10-07T09:00:00+07:00'
        valid_until: '2026-10-07T10:30:00+07:00'
        revoked_at: null
    GrantRevoked:
      value:
        grant_id: gr_8Fv2Lq9Xk3Tw7Mz1Rb5Nc4
        booking_reference: BK-7KQ2M9
        status: revoked
        condition: null
        ticket_code: H7K3-9QXA
        ticket_url: http://localhost:8003/t/q3Vt9XbLk2Rm8PzW4nHc7A
        space_id: 1
        space_name: Meeting Room A
        valid_from: '2026-10-07T09:00:00+07:00'
        valid_until: '2026-10-07T10:30:00+07:00'
        revoked_at: '2026-10-05T11:00:00+07:00'
    GrantTombstone:
      value:
        grant_id: gr_4Hs9Pd2Wm6Qx8Kn3Tf7Lz1
        booking_reference: BK-P4W6RC
        status: revoked
        condition: null
        ticket_code: null
        ticket_url: null
        space_id: null
        space_name: null
        valid_from: null
        valid_until: null
        revoked_at: '2026-10-05T10:05:00+07:00'